writeLong should return int not long

+++ protocol/src/main/java/pl/grzeslowski/jsupla/protocol/api/encoders/PrimitiveEncoder.java
@@ -6,7 +6,7 @@ public interface PrimitiveEncoder {
 
     int writeUnsignedInteger(long value, byte[] bytes, int offset);
 
-    long writeLong(long value, byte[] bytes, int offset);
+    int writeLong(long value, byte[] bytes, int offset);
 
     int writeByte(byte value, byte[] bytes, int offset);
